Move events and errors stories => examples
