+++ examples/official-storybook/stories/core/errors.examples.js
@@ -1,0 +1,18 @@
+export default {
+  title: 'Core|Errors',
+};
+
+export const exception = () => {
+  throw new Error('error');
+};
+exception.title = 'story throws exception';
+exception.parameters = {
+  storyshots: { disabled: true },
+};
+
+export const errors = () => null;
+errors.title = 'story errors';
+errors.parameters = {
+  notes: 'Story does not return something react can render',
+  storyshots: { disabled: true },
+};
